+++ packages/jaspr/lib/src/client/run_app.dart
@@ -3,5 +3,7 @@ import 'client_binding.dart';
 
 /// Main entry point for the browser app.
 ClientAppBinding runApp(Component app, {String attachTo = 'body'}) {
-  return ClientAppBinding()..attachRootComponent(app, attachTo: attachTo);
+  final binding = ClientAppBinding();
+  Future.microtask(() => binding.attachRootComponent(app, attachTo: attachTo));
+  return binding;
 }