+++ packages/jaspr/lib/src/framework/render_object.dart
@@ -95,7 +95,7 @@ mixin RenderObjectElement on Element {
 
   @override
   void detachRenderObject() {
-    final parent = _parentRenderObjectElement?.renderObject;
+    final parent = renderObject.parent;
     if (parent != null) {
       parent.remove(renderObject);
       assert(renderObject.parent == null);
@@ -106,6 +106,15 @@ mixin RenderObjectElement on Element {
   @override
   void updateSlot(ElementSlot newSlot) {
     super.updateSlot(newSlot);
-    attachRenderObject();
+    if (!owner._isReload) {
+      attachRenderObject();
+    }
+  }
+
+  @override
+  void onReload() {
+    _renderObject = createRenderObject();
+    _attached = false;
+    super.onReload();
   }
 }