+++ packages/jaspr/lib/src/framework/stateful_component.dart
@@ -578,6 +578,11 @@ abstract class State<T extends StatefulComponent> {
   @protected
   @mustCallSuper
   void didChangeDependencies() {}
+
+  /// Called when the page is reloaded.
+  @protected
+  @mustCallSuper
+  void onReload() {}
 }
 
 /// Mixin on [State] that preloads state on the server
@@ -751,5 +756,11 @@ class StatefulElement extends BuildableElement {
   void didChangeDependencies() {
     super.didChangeDependencies();
     _didChangeDependencies = true;
+  }
+
+  @override
+  void onReload() {
+    state.onReload();
+    super.onReload();
   }
 }